first pass on rest api
Been a long time since ive used go... or docker for that matter. Realized i was in a groove of coding and testing and not committing along the way, should do so now!

Working as is, but want to do some cleanup and safeties, then need to work on testing, and then remind myself how to deal with docker for this